pyproject.toml
[tool.poetry]
name = "iteration-report"
version = "0.1.0"
description = ""
authors = ["ISON <azzgo@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.pdm.scripts]
start = "streamlit run main.py --server.headless true --server.address 0.0.0.0 --server.port 3001"

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "notebook<8.0.0,>=7.1.3",
]

[tool.pdm.build]
includes = []
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[project]
authors = [
    {name = "ISON", email = "azzgo@users.noreply.github.com"},
]
requires-python = "<4.0,>=3.11"
dependencies = [
    "python-dotenv<2.0.0,>=1.0.1",
    "httpx<1.0.0,>=0.27.0",
    "streamlit<2.0.0,>=1.33.0",
    "watchdog<5.0.0,>=4.0.0",
    "streamlit-echarts<1.0.0,>=0.4.0",
]
name = "iteration-report"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false
